+++ parity-top/src/main/java/org/jvirtanen/parity/top/Orders.java
@@ -34,7 +34,7 @@ class Orders {
         return levels.get(levels.firstLongKey());
     }
 
-    public Order add(long price, int size) {
+    public Order add(long price, long size) {
         Level level = levels.get(price);
         if (level == null) {
             level = new Level(this, price);
